Update faq.ejs
RULES FOR ELEPHANT CHESS !
1. The elephant piece can NEVER be Captured NOR can it Capture any chess piece.

2. The elephant piece is obtained by a player who captures an opponent's piece.

3. The elephant piece can be placed on ANY empty square.

4. The square the elephant piece currently resides becomes IMPASSABLE to both players.

DOMINATION or CATCH and RELEASE!
Domination mode: A player who currently dominates(controls) the elephant piece can choose to move it OR a regular chess piece when it's their turn but cannot move the elephant and a regular chess piece simultaneously(at the same time).

(Exception to this rule Only applies to the first player to capture a chess piece!)

Catch and Release mode: A player who captures an opponent's piece can move the elephant simultaneously(at the same time) upon capturing their opponent's chess piece.

Addendum: The elephant piece is used to hinder/disrupt an opponent's strategies during the course of the game. It can be used offensively as well as defensively.

We Hope You Will Enjoy Elephant Chess!